extensions/middleware/vector/requirements.txt
# Required dependencies
weaviate-client>=4.0.0
numpy>=1.24.0
pytest>=7.3.1
pytest-asyncio>=0.21.0

# Optional dependencies for embeddings (uncomment as needed)
# sentence-transformers>=2.2.0
# openai>=1.3.0
# cohere>=4.32 